python/packages/kagent-adk/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "kagent-adk"
version = "0.3.0"
description = "kagent-adk is an sdk for integrating adk agents with kagent"
readme = "README.md"
requires-python = ">=3.12.11"
dependencies = [
  "openai>=1.72.0",
  "mcp>=1.8.1",
  "protobuf>=6.31.1",
  "opentelemetry-api>=1.32.0",
  "opentelemetry-sdk>=1.32.0",
  "opentelemetry-exporter-otlp-proto-grpc>=1.32.0",
  "opentelemetry-instrumentation-openai>= 0.39.0",
  "opentelemetry-instrumentation-httpx >= 0.52.0",
  "anthropic[vertex]>=0.49.0",
  "fastapi>=0.103.1",
  "litellm>=1.74.3",
  "google-adk>=1.7.0",
  "google-genai>=1.18.0",
  "google-auth>=2.40.2",
  "httpx>=0.25.0",
  "pydantic>=2.5.0",
  "typing-extensions>=4.8.0",
  "jsonref>=1.1.0",
  "a2a-sdk>=0.2.16",
]

[project.optional-dependencies]
test = [
  "pytest>=8.3.5",
  "pytest-asyncio>=0.25.3",
]

[tool.hatch.build.targets.wheel]
packages = ["src/kagent_adk"]

[tool.ruff]
extend = "../../pyproject.toml"